xplat: reduce source changes between Static and Shared Libraries.
This greatly reduces the complexity of switching between static and shared
libraries by converting most of the static libraries to Shared libraries.
Now it is possible to test a static build of chakracore by simply linking
aginst libChakraCoreStatic. Also, this greatly reduces compile time when
changing between the two since the changes between static and shared
libraries has been isolated to within the bin folder.

v2: Add Compile Fixes for MSVC/OSX, and fix native tests on ubuntu.
