+++ bin/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory (GCStress)
+add_subdirectory (ChakraCore)
 add_subdirectory (ch)
-if(NOT STATIC_LIBRARY)
-    add_subdirectory (ChakraCore)
-endif()
+
+
