+++ bin/ChakraCore/CMakeLists.txt
@@ -1,8 +1,38 @@
+
+
+add_library (ChakraCoreStatic STATIC 
+  ChakraCoreCommon.cpp
+  ConfigParserExternals.cpp
+  TestHooks.cpp
+  ${Chakra.Jsrt}
+  )
+
 add_library (ChakraCore SHARED
+  ChakraCoreCommon.cpp
   ChakraCoreDllFunc.cpp
   ConfigParserExternals.cpp
   TestHooks.cpp
-)
+  )
+
+target_compile_options(ChakraCoreStatic PRIVATE "-fPIC")
+target_compile_definitions(ChakraCoreStatic
+  PUBLIC
+    -DCHAKRA_STATIC_LIBRARY=1
+  )
+CHAKRACORE_ADD_JSRT_OBJECTS(ChakraCore)
+CHAKRACORE_ADD_JSRT_OBJECTS(ChakraCoreStatic)
+
+target_include_directories (
+  ChakraCoreStatic PUBLIC
+  ${CMAKE_CURRENT_SOURCE_DIR}
+  ${CHAKRACORE_SOURCE_DIR}/lib/Common
+  ${CHAKRACORE_SOURCE_DIR}/lib/Runtime
+  ${CHAKRACORE_SOURCE_DIR}/lib/Parser
+  ${CHAKRACORE_SOURCE_DIR}/lib/Jsrt
+  ${CHAKRACORE_SOURCE_DIR}/lib/JITIDL
+  ${CHAKRACORE_SOURCE_DIR}/lib/Backend
+  ${CHAKRACORE_SOURCE_DIR}/lib/Runtime/ByteCode/
+  )
 
 target_include_directories (
   ChakraCore PUBLIC
@@ -11,6 +41,9 @@ target_include_directories (
   ${CHAKRACORE_SOURCE_DIR}/lib/Runtime
   ${CHAKRACORE_SOURCE_DIR}/lib/Parser
   ${CHAKRACORE_SOURCE_DIR}/lib/Jsrt
+  ${CHAKRACORE_SOURCE_DIR}/lib/JITIDL
+  ${CHAKRACORE_SOURCE_DIR}/lib/Backend
+  ${CHAKRACORE_SOURCE_DIR}/lib/Runtime/ByteCode/
   )
 
 #
@@ -39,9 +72,9 @@ endif()
 set(lib_target "${lib_target}"
   -Wl,-undefined,error
   ${LINKER_START_GROUP}
-  Chakra.Pal
-  Chakra.Common.Core
-  Chakra.Jsrt
+  #Chakra.Pal
+  #Chakra.Common.Core
+  #Chakra.Jsrt
   ${LINKER_END_GROUP}
   pthread
   stdc++
@@ -49,16 +82,36 @@ set(lib_target "${lib_target}"
   ${ICULIB}
   )
 
-if(CMAKE_SYSTEM_NAME STREQUAL Linux)
+if(CMAKE_SYSTEM_NAME STREQUAL Darwin)
   set(lib_target "${lib_target}"
-    -Wl,--version-script=${CMAKE_CURRENT_SOURCE_DIR}/libChakraCoreLib.version
-    )
+    "-framework CoreFoundation"
+    "-framework Security"
+  )
+elseif(CMAKE_SYSTEM_NAME STREQUAL Linux)
+  if(CC_TARGETS_AMD64)
+      set(lib_target "${lib_target}"
+        unwind-x86_64
+      )
+  endif()
+
+  set(lib_target "${lib_target}"
+    gcc_s
+    rt
+    unwind
+    unwind-generic
+  )
 endif()
 
 if(CC_TARGETS_X86)
   set(lib_target "${lib_target} -m32")
 endif()
+target_link_libraries (ChakraCoreStatic ${lib_target})
 
+if(CMAKE_SYSTEM_NAME STREQUAL Linux)
+  set(lib_target "${lib_target}"
+    -Wl,--version-script=${CMAKE_CURRENT_SOURCE_DIR}/libChakraCoreLib.version
+    )
+endif()
 target_link_libraries (ChakraCore ${lib_target})
 
 if(NOT CC_XCODE_PROJECT)
@@ -72,6 +125,9 @@ if(NOT CC_XCODE_PROJECT)
   add_custom_command(TARGET ChakraCore POST_BUILD
     COMMAND ${CMAKE_COMMAND} -E copy_if_different
     "${CHAKRACORE_BINARY_DIR}/bin/ChakraCore/libChakraCore.${CC_LIB_EXT}"
+    ${CHAKRACORE_BINARY_DIR}/
+    COMMAND ${CMAKE_COMMAND} -E copy_if_different
+    "${CHAKRACORE_BINARY_DIR}/bin/ChakraCore/ChakraCoreStatic.a"
     ${CHAKRACORE_BINARY_DIR}/
     )
 endif()
