+++ lib/Common/Core/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library (Chakra.Common.Core STATIC
+add_library (Chakra.Common.Core OBJECT
     Api.cpp
     BinaryFeatureControl.cpp
     CmdParser.cpp
