+++ lib/Jsrt/CMakeLists.txt
@@ -1,8 +1,5 @@
-if(BuildJIT)
-    set(chakra_backend_objects $<TARGET_OBJECTS:Chakra.Backend>)
-endif()
 
-add_library (Chakra.Jsrt STATIC
+add_library (Chakra.Jsrt.object OBJECT
     Jsrt.cpp
     JsrtDebugUtils.cpp
     JsrtDebugManager.cpp
@@ -17,12 +14,27 @@ add_library (Chakra.Jsrt STATIC
     JsrtRuntime.cpp
     JsrtSourceHolder.cpp
     JsrtThreadService.cpp
+    )
+
+macro(CHAKRACORE_ADD_JSRT_OBJECTS _TARGET)
+  target_sources(${_TARGET} PRIVATE
+    $<TARGET_OBJECTS:Chakra.Pal>
+    $<TARGET_OBJECTS:Chakra.Common.Core>
+    $<TARGET_OBJECTS:Chakra.Common.Memory>
+    $<TARGET_OBJECTS:Chakra.Common.Exceptions>
+    $<TARGET_OBJECTS:Chakra.Runtime.PlatformAgnostic>
     $<TARGET_OBJECTS:Chakra.Jsrt.Core>
-    ${chakra_backend_objects}
+    )
+  if(TARGET Chakra.Backend)
+    target_sources(${_TARGET} PRIVATE
+      $<TARGET_OBJECTS:Chakra.Backend>
+      )
+  endif()
+  target_sources(${_TARGET} PRIVATE  
 #   Do not take this in. We need to control the
 #   linker order because of global constructors
 #   and cross dependencies among them
-#   $<TARGET_OBJECTS:Chakra.Common.Core>
+    $<TARGET_OBJECTS:Chakra.Common.Core>
     $<TARGET_OBJECTS:Chakra.Common.Common>
     $<TARGET_OBJECTS:Chakra.Common.Codex>
     $<TARGET_OBJECTS:Chakra.Common.DataStructures>
@@ -38,12 +50,15 @@ add_library (Chakra.Jsrt STATIC
     $<TARGET_OBJECTS:Chakra.Runtime.Types>
     $<TARGET_OBJECTS:Chakra.Runtime.PlatformAgnostic>
     $<TARGET_OBJECTS:Chakra.Parser>
+    $<TARGET_OBJECTS:Chakra.Jsrt.object>
     )
+endmacro(CHAKRACORE_ADD_JSRT_OBJECTS)
+
 
 add_subdirectory(Core)
 
 target_include_directories (
-    Chakra.Jsrt PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}
+    Chakra.Jsrt.object PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}
     ../Backend
     ../JITIDL
     ../Runtime
